reduce frequency of label presence checks
