+++ src/index.ts
@@ -1,7 +1,6 @@
 import { Probot } from 'probot';
 
 import Handlers from './handlers/index.js';
-import { labels } from './labels.js';
 
 export default (app: Probot) => {
   /*
@@ -16,48 +15,6 @@ export default (app: Probot) => {
     });
   });
   */
-
-  app.on(
-    ['pull_request', 'pull_request_review', 'check_suite'],
-    async context => {
-      const owner = context.payload.repository.owner.login;
-      const repo = context.payload.repository.name;
-
-      const existingLabels = (
-        await context.octokit.issues.listLabelsForRepo({
-          owner,
-          repo,
-        })
-      ).data;
-
-      for (const label of Object.values(labels)) {
-        const existing = existingLabels.find(l => l.name === label.name);
-        const color = label.color.substring(1).toUpperCase();
-
-        if (!existing) {
-          await context.octokit.issues.createLabel({
-            owner,
-            repo,
-            name: label.name,
-            color,
-          });
-
-          continue;
-        }
-
-        if (existing.color.toUpperCase() !== color) {
-          await context.octokit.issues.updateLabel({
-            owner,
-            repo,
-            name: label.name,
-            color,
-          });
-
-          continue;
-        }
-      }
-    },
-  );
 
   app.log.info('probot loaded');
   Handlers(app);
